pyproject.toml
[tool.poetry]
name = "nisman"
version = "0.1.0"
description = "Proyecto Nisman"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.11"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
google-cloud-storage = "^2.14.0"
google-cloud-pubsub = "^2.18.4"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
